+++ class-components/src/pages/DetailPage/DetailPage.test.tsx
