+++ class-components/src/pages/MainPage/MainPage.test.tsx
