+++ class-components/src/pages/MainPage/MainPage.tsx
@@ -1,12 +1,12 @@
 import { ChangeEvent, useEffect, useState } from 'react';
-import Input, { InputProps } from '../components/Input';
-import SearchButton, { ButtonProps } from '../components/SearchButton';
-import fetchData, { defaultURL } from '../utils/api';
-import CardsContainer from '../components/CardsContainer';
-import ErrorBoundary from '../components/ErrorBoundary';
-import ErrorButton from '../components/ErrorButton';
-import useLocalStorage from '../utils/localStorage';
-import PaginationButton from '../components/PaginationButton';
+import Input, { InputProps } from '../../components/Input';
+import SearchButton, { ButtonProps } from '../../components/SearchButton';
+import fetchData, { defaultURL } from '../../utils/api';
+import CardsContainer from '../../components/Cards/CardsContainer';
+import ErrorBoundary from '../../components/ErrorJHandling/ErrorBoundary';
+import ErrorButton from '../../components/ErrorJHandling/ErrorButton';
+import useLocalStorage from '../../utils/localStorage';
+import PaginationButton from '../../components/PaginationButton';
 import { Outlet, useNavigate } from 'react-router-dom';
 import { useSearchParams } from 'react-router-dom';
